+++ cpp/src/arrow/python/arrow_to_pandas.cc
@@ -36,7 +36,7 @@
 #include "arrow/type_fwd.h"
 #include "arrow/type_traits.h"
 #include "arrow/util/bit-util.h"
-#include "arrow/util/int128.h"
+#include "arrow/util/decimal.h"
 #include "arrow/util/logging.h"
 #include "arrow/util/macros.h"
 #include "arrow/util/parallel.h"
@@ -605,7 +605,7 @@ static Status ConvertTimes(PandasOptions options, const ChunkedArray& data,
 static Status RawDecimalToString(const uint8_t* bytes, int precision, int scale,
                                  std::string* result) {
   DCHECK_NE(result, nullptr);
-  Int128 decimal(bytes);
+  Decimal128 decimal(bytes);
   *result = decimal.ToString(precision, scale);
   return Status::OK();
 }
